add fields.py
